When clicked on the button below the file list, call the new task method explained above, and redirect the user to the “Laboratory” tab.:
Code implemented but GET request no working

Call every second the method to know what is the current state. Keep it in a state variable (React’s useState([])):

same as above

Show what is the current state in the user interface in the Laboratory tab.

If the user changes tabs (going to Load files or intro), cancel the current task if there is a task ongoing (i.e., not in error / completed / etc.)
Function code, need to refactor nav bar